requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
Pillow==10.4.0
Werkzeug==3.1.4
